Add files via upload
The aim of the system was to enable the doctors and the responsible nurse to submit orders for personal and general use materials respectively, through the Procurement Department. Utilizing the capabilities of the Bonita Studio, we have sought to create a smooth flow of efficiency for efficiency and ensures timely access to basic health care products.
In order to deal with possible deficiencies, our system incorporates a smart mechanism. If the order data are not available, the supply section activates a procedure produced by a responsible member a list of inadequate materials. This list is subsequently promoted to reliable external supply partners, ensuring uninterrupted access to critical supplies.